backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg[binary]>=3.2.4
psycopg_pool>=3.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.0
pydantic>=2.0.0
slowapi==0.1.9
email-validator>=2.0.0
redis>=5.0.0
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[fastapi]>=1.40.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

